Fix GPU stochasticity: position randomization and type mismatch
Cherry-picked from adacs-gpu-debug branch:
- Fix halo_coords type mismatch (int* -> float*)
- Implement position randomization for progenitor halos

The GPU kernel was copying the descendant's coordinates to ALL
progenitors. The CPU version uses random_point_in_sphere to place
each progenitor randomly within the descendant's Lagrangian sphere.

New device functions:
- device_MtoR: Convert halo mass to Lagrangian radius
- device_random_point_in_sphere: Generate random point in sphere
- device_wrap_position: Wrap coordinates to periodic boundaries

Fixes low correlations in discrete halo scripts.

+++ src/py21cmfast/src/Stochasticity.c
@@ -989,7 +989,7 @@ int sample_halo_progenitors(gsl_rng **rng_arr, double z_in, double z_out, HaloFi
         float *halo_star_rng = halofield_in->star_rng;
         float *halo_sfr_rng = halofield_in->sfr_rng;
         float *halo_xray_rng = halofield_in->xray_rng;
-        int *halo_c = halofield_in->halo_coords;
+        float *halo_c = halofield_in->halo_coords;  // FIX: was int*, but halo_coords is float*
 
         updateHaloOut(halo_m, halo_star_rng, halo_sfr_rng, halo_xray_rng, halo_c, nhalo_in,
                       sigma_y_arr, sigma_bin, x_min, x_width, d_hs_constants, arraysize_total,
